Make search bar more extensible and factor out utils
